==> 40_carinspricing_modelcheck.Rmd <==
---
title: "GLMs for Car Insurance Pricing: Model Assessment"
author: "Mick Cooney <mickcooney@gmail.com>"
date: "13 July 2016"
output:
  html_document:
    fig_caption: yes
    theme: spacelab #sandstone #spacelab #flatly
    highlight: pygments
    number_sections: TRUE
    toc: TRUE
    toc_depth: 2
    toc_float:
      smooth_scroll: FALSE
  pdf_document: default
---



```{r knit_opts, include = FALSE}
rm(list = ls())

knitr::opts_chunk$set(tidy       = FALSE
                     ,cache      = FALSE
                     ,fig.height =  8
                     ,fig.width  = 11
                     )

library(tidyverse)
library(data.table)
library(dtplyr)

library(feather)
library(poweRlaw)


options(width            = 90L)
options(stringsAsFactors = FALSE)

options(datatable.print.nrows      = 10L)
options(datatable.prettyprint.char = 80L)

source("custom_functions.R")
```


This document is part of a series investigating the use of generalised linear
models to price car insurance. This document assesses the pricing models built
previously and uses MonteCarlo simulation to discover the range of possible
outcomes that are possible assuming we get the data correct. The main focus is
to get assess the distribution of possible outcomes.



# Set Parameters and Load Data

Before we begin we configure some parameters needed for the
simulations.

```{r config_params, echo=TRUE}
set.seed(242)

n_sim <- 500

largeloss_threshold <- 5000
largeloss_sf        <- 2.72
largeloss_rate      <- 0.001588
```

We load the various datasets, the model fits and the price calculation.

```{r load_data, echo=TRUE}
policy_dt <- read_feather("data/policy_data.feather")
claim_dt  <- read_feather("data/claim_data.feather")

trainpolicy_dt <- read_feather("data/train_data.feather")
validpolicy_dt <- read_feather("data/valid_data.feather")
testpolicy_dt  <- read_feather("data/test_data.feather")


setDT(trainpolicy_dt)
setDT(validpolicy_dt)
setDT(testpolicy_dt)

glimpse(trainpolicy_dt)
glimpse(validpolicy_dt)
glimpse(testpolicy_dt)

claimrate_model_glm <- readRDS("data/claim_model.rds")
claimsize_model_glm <- readRDS("data/size_model.rds")

premium_quoter <- readRDS("data/premium_quoter.rds")
```

In this document we focus assessing our pricing model. No model is
perfect, and it is important to understand how a pricing model may
fail and what risks are posed.

To start with, we make a very strong and generous assumption: we
assume our pricing model is correct and the risks assumed by the
business are represented correctly by the models.

This assumption is absurd, we already know the model has flaws, and
spent part of the last document working on them. That said, it makes
for a easy and frictionless starting point, and should help illustrate
the extent of tail risk we have in the model.

# Evaluating the Model with Same Parameters

The idea is to use the same parameters used to price the policies to
simulate claim data. For each simulation iteration and policy, we
perform a number of steps:

* Use the Poisson model to estimate claim counts for the policy
* Use the Gamma model to estimate the size of the claims
* Use the large loss claim frequency to estimate the count of large
  claims
* Use the power law calculation to simulate the size of the large
  claims

Note that for performance reasons the logical steps outlined above may
be implemented differently without changing the outcome of the
simulation overall.

Having done the above for every policy and claim, we run multiple
MonteCarlo simulations to estimate the distribution of possible
outcomes *conditional on the model being correct*.

This approach may seem pointless - using a model to validate itself -
but it is not. It gives us a sense of the best case scenario from a
probabilistic perspective.


## Converting GLM Output to Distribution Parameters

To calculate simulations we require, we need to convert the output of
GLM regressions to parameters for the underlying distributions. GLMs
fit the data to produce the mean and dispersion of the data - perfect
for the use of a normal distribution as it is parameterised by the
mean and variance - but other distributions use different parameters.

The Poisson distribution has a single parameter, the mean, so we can
use that output directly if we wish to simulate from a Poisson.

The Gamma distribution has a few choices for parameterisation, the
most common of which is to use the shape, $k$ and the rate,
$\beta$. The shape is the reciprocal of the dispersion in the dataset,
and the rate is given by

$$ \beta = \frac{k}{\mu}. $$

With these transformations we can use the model fits to produce data
in our simulations.


## Assessing the Model with Training Data

```{r model_assess_simple, echo=TRUE}

simulator_creator <- create_assessment_simulator(claimrate_model_glm
                                                ,claimsize_model_glm
                                                ,largeloss_threshold
                                                ,largeloss_rate
                                                ,largeloss_sf)
```

With this infrastructure in place we are now in position to assess the
model using the training dataset.

For each MonteCarlo iteration we simulate a claim count for each
policy based on the Poisson model we fit (we exclude large claims for
now). For each policy with at least one claim, we sample a claim
amount using the gamma distribution.

With the 'usual' losses simulated, we then sample a total number of
larger claims using a Poisson distribution with the large claim rate
and then use that as the number of samples we draw from the power law
distribution. The sum of these two claim amounts are the total claims
on the book for that iteration.

We perform the above simulation for a large number of simulations.

```{r assess_training_data, echo=TRUE, cache=TRUE}
use_dt <- trainpolicy_dt

run_mc_train_sim <- simulator_creator(use_dt)

sim_train_losses <- replicate(n_sim, run_mc_train_sim())

total_riskprem   <- sum(use_dt$risk_premium)
received_premium <- sum(use_dt$price_quote)
```

Having ran the simulations we now plot the distribution of the
simulated claim payments against the premium received.

```{r assess_training_data_plot, echo=TRUE}
ggplot() +
    stat_density(aes(x = sim_train_losses['total_claims',]), geom = 'line') +
    geom_vline(aes(xintercept = received_premium), colour = 'red') +
    geom_vline(aes(xintercept = total_riskprem),   colour = 'blue') +
    scale_x_continuous(labels = scales::dollar)
```

This looks poor, much worse than we would anticipate given that we are
reusing the parameters in the simulation. The risk premium receive is
way below the amount of claims paid out. Even the total premium
received - which should cover the additional costs of running the
business and profit margin - do not cover the losses. Why do the
outputs look so bad?

Is it possible that using the expectation for the Gamma distribution
for the pricing is not a good idea due to the tail? We can check this
by suppressing the simulation of the claim amount and instead using a
fixed amount for each claim and seeing how this looks.

```{r assess_training_data_fixedclaim, echo=TRUE, cache=TRUE}
run_mc_train_fixedclaim_sim <- simulator_creator(use_dt
                                                ,simulate_claim_size = FALSE)

sim_train_fixedclaim_losses <- replicate(n_sim, run_mc_train_fixedclaim_sim())

total_riskprem   <- sum(use_dt$risk_premium)
received_premium <- sum(use_dt$price_quote)
```

```{r assess_training_data_fixedclaim_plot, echo=TRUE}
ggplot() +
    stat_density(aes(x = sim_train_fixedclaim_losses['total_claims',]), geom = 'line') +
    geom_vline(aes(xintercept = received_premium), colour = 'red') +
    geom_vline(aes(xintercept = total_riskprem),   colour = 'blue') +
    scale_x_continuous(labels = scales::dollar)
```



```{r assess_validation_data, echo=TRUE, cache=TRUE}
use_dt <- validpolicy_dt

run_mc_valid_sim <- simulator_creator(use_dt)

sim_valid_losses <- replicate(n_sim, run_mc_valid_sim())

total_riskprem   <- sum(use_dt$risk_premium)
received_premium <- sum(use_dt$price_quote)
```

```{r assess_validation_data_plot, echo=TRUE}
ggplot() +
    stat_density(aes(x = sim_valid_losses['total_claims',]), geom = 'line') +
    geom_vline(aes(xintercept = received_premium), colour = 'red') +
    geom_vline(aes(xintercept = total_riskprem),   colour = 'blue') +
    scale_x_continuous(labels = scales::dollar)
```


# R Environment

```{r show_session_info, echo=TRUE}
sessioninfo::session_info()
```
